requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
braceexpand==0.1.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
cytoolz==1.0.1
datasets==3.5.0
decorator==5.2.1
dill==0.3.8
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
executing==2.2.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.12.0
grpcio==1.71.0
huggingface-hub==0.30.2
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
intervaltree==3.1.0
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiwer==3.1.0
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.8
lazy_loader==0.4
lhotse==1.27.0
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.14.3
lilcom==1.8.1
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
nemo-toolkit==2.2.1
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
omegaconf==2.3.0
onnx==1.17.0
optuna==4.3.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
primePy==1.3
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==3.20.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==19.0.1
pycparser==2.22
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
SQLAlchemy==2.0.40
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.21.1
toolz==1.0.0
torch==2.7.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.0
torchmetrics==1.7.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.15.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
webdataset==0.2.111
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
